feat(build): offer the flakiness prompt on a flaky snapshot
The snapshot sidebar already says a capture keeps moving on its own — how many
of the last seven days' builds produced the same change, a flakiness score the
gauge paints red — and offered nothing to do about it but ignore it. Test
Insights now hands out the prompt the test page has, on the snapshots Argos
already calls flaky: a change that keeps coming back, or a test the gauge does
not paint green.

The prompt is the test's, not the capture's, and unchanged: a flake is what a
test does over time, so no single screenshot of it is the right one. It moves
to `containers/Test` since two pages hand it out now, which is where
`Flakiness.ts` already lives for the same reason.

Prompt buttons also name the action rather than the destination — "Fix
flakiness" instead of "Open in Claude". The agent's mark carries where it goes,
and the tooltip still spells it out.
